+++ workspaces/redhat-argocd/plugins/argocd/src/utils/__tests__/utils.test.ts
@@ -14,7 +14,11 @@
  * limitations under the License.
  */
 import { mockApplication, mockEntity } from '../../../dev/__data__';
-import { Application, History, Status } from '../../types/application';
+import {
+  Application,
+  History,
+  Status,
+} from '@backstage-community/plugin-redhat-argocd-common';
 import {
   ArgoCdLabels,
   getAppOperationState,