+++ workspaces/redhat-argocd/plugins/argocd/CHANGELOG.md
@@ -1,5 +1,16 @@
 ## @backstage-community/plugin-redhat-argocd
 
+## 1.16.0
+
+### Minor Changes
+
+- bba0c01: Add multi-source application support and move application types to plugin-redhat-argocd-common package.
+
+### Patch Changes
+
+- Updated dependencies [bba0c01]
+  - @backstage-community/plugin-redhat-argocd-common@1.3.0
+
 ## 1.15.0
 
 ### Minor Changes